Add zIndex parameter to Draw and InitUniformBuffer methods
Added zIndex as a parameter in Draw and InitUniformBuffer methods across various files to improve depth calculations. This allows us to adjust the positioning of objects along z-axis, which can enhance depth perception in the 3D space. zIndex is now an integral part of every draw call and transformation, improving the flexibility in handling object's depth.

+++ include/Util/GameObject.hpp
@@ -13,7 +13,7 @@ namespace Util {
 
         GameObject() = default;
 
-        GameObject(std::unique_ptr<Core::Drawable> drawable, const int zIndex,
+        GameObject(std::unique_ptr<Core::Drawable> drawable, const float zIndex,
                    const std::vector<std::shared_ptr<GameObject>> &children =
                    std::vector<std::shared_ptr<GameObject>>())
                 : m_Drawable(std::move(drawable)),
